refactor: change src/dht file structure
